+++ packages/components/src/Popover/Popover.tsx
@@ -65,7 +65,7 @@ export interface UsePopoverProps {
    */
   isOpen?: boolean
   /**
-   * Specify a callback to be called each time this Modal is closed
+   * Specify a callback to be called each time this Popover is closed
    */
   onClose?: () => void
   /**
@@ -84,9 +84,9 @@ export interface UsePopoverProps {
   content: ReactNode
 
   /**
-   * Specify a callback to be called before trying to close the Modal. This allows for
+   * Specify a callback to be called before trying to close the Popover. This allows for
    * use-cases where the user might lose work (think common "Save before closing warning" type flow)
-   * Specify a callback to be called each time this Modal is closed
+   * Specify a callback to be called each time this Popover is closed
    */
   canClose?: () => boolean
 
@@ -158,7 +158,7 @@ export interface PopoverProps extends UsePopoverProps {
   /**
    * Component to wrap. The HOC will listen for mouse events on this
    * component, maintain the state of isOpen accordingly, and pass that state into
-   * the modal renderProp.
+   * the Popover renderProp.
    */
   children: ReactNode | PopoverRenderProp